TUSC-142 prework: Add routes (#30)
Add routes

This adds the routes the application needs to run. Specfically they are:

* gold-images
* cloud-accounts
* marketplace-purchases

It also adds some constants and helpers for routing, so that a route
changing does not require lots of manual changes.

The HelloCloudInventoryPage has also been updated to not have random
padding on the left side, and it renders at every route now.