TUSC-341: Add useQueryParamInformed<State|Atom> (#36)
* Add useQueryInformed<State|Atom>

This adds two hooks, `useQueryInformedState` and `useQueryInformedAtom`
to allow for convenient manipulation of query parameters. These two
functions are interchangeable with their native counterparts, just
requiring the addition of a `key` for which to look for, and set, query
params with.

* Handle non-json query param and use safer decode/encode methods

This updates to handle an error on non-json in a query param, and ignore
it if it fails to parse. It also switch to the `decodeURIComponent` and
`encodeURIComponent` as well to ensure safely escaped characters in
query parameters.